+++ packages/dataviews/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - DataForm: rename "Select item" to "Select" and disable empty `"Select"` option when the field is required and other option has been selected. [#72042](https://github.com/WordPress/gutenberg/pull/72042)
 - Move search icon in search fields to prefix position ([#71984](https://github.com/WordPress/gutenberg/pull/71984)).
+- Flip search icons depending on placement ([#72070](https://github.com/WordPress/gutenberg/pull/72070)).
 - DataViews: Improve renderItemLink event propagation handling. ([#72081](https://github.com/WordPress/gutenberg/pull/72081)).
 
 ### Breaking changes